RESISTE-118 Transfered all of the parsing logic into RssEntry#setContents(). Created an RssEntryTest unit test.
